feat: 添加分数计分和显示功能
1. 新增ScoreArea标签与碰撞检测区域，实现 passing pipe 计分逻辑
2. 重构分数相关变量为int类型，添加当前分数和最高分的更新与显示方法
3. 完善管道预制体的分数区域配置，动态适配管道间隙尺寸
4. 在主场景添加分数和最高分的UI文本并绑定逻辑

+++ Assets/PhoebeRunFast/Scripts/Entity/FlyBirdEntity.cs
@@ -27,7 +27,7 @@ public class FlyBirdEntity : IEntity
 	/// </summary>
 	public float globalPipeSpeed = 10f;
 
-	public float maxScore = 0;
+	public int maxScore = 0;
 
-	public float currentScore = 0;
+	public int currentScore = 0;
 }
